+++ api/package.json
@@ -92,6 +92,7 @@
     "lineage",
     "message-utils",
     "phone-number",
+    "purging-utils",
     "registration-utils",
     "search",
     "server-checks",
