+++ shared-libs/purging-utils/package.json
@@ -1,0 +1,16 @@
+{
+  "name": "@medic/purging-utils",
+  "version": "1.0.0",
+  "description": "Purging common utility functions",
+  "main": "src/index.js",
+  "scripts": {
+    "test": "mocha ./test"
+  },
+  "author": "",
+  "license": "Apache-2.0",
+  "devDependencies": {
+    "chai": "^4.2.0",
+    "mocha": "^5.2.0",
+    "sinon": "^7.1.1"
+  }
+}
